requirements.txt
python==3.6.12
scipy==1.1.0
opencv_python==4.4.0.44
matplotlib==3.2.2
numpy==1.19.1
pandas==1.1.1
Pillow==7.2.0
scikit_learn==0.23.2
fastai==2.0.15
pytorch==1.6.0
